backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
beautifulsoup4
langchain
faiss-cpu
google-genai
